make Mandate report abstract
Make the Report tree a little cleaner by emphasising that the 'Mandate'
class is abstract, and can be used by either the single-member Office
report, or the multi-member Term report, and pushing all the 'look at
predecessor and successor' logic specifically into the Office report.
